[Loom] Isolate source callable boundaries
Give source callable ABI lowering a documented component that owns
signature mapping, Low callable creation, direct and resource argument
binding, predicate translation, and bodyless import lowering. Keep CFG
and body emission in the orchestrator.

Build target callables without source predicates, then remap predicate
value references onto the new boundary and register their attribute uses
before erasing the source callable. This prevents imported declarations
from retaining stale source value IDs and keeps definition use
bookkeeping consistent.
