+++ loom/src/loom/codegen/low/lower/BUILD.bazel
@@ -30,6 +30,7 @@ loom_cc_library(
     name = "lower",
     srcs = [
         "contract_query.c",
+        "function_boundary.c",
         "lower.c",
         "lower_context.c",
         "lower_policy.c",
@@ -44,6 +45,7 @@ loom_cc_library(
     ],
     hdrs = [
         "contract_query.h",
+        "function_boundary.h",
         "lower.h",
         "lower_internal.h",
         "lower_report.h",
@@ -164,6 +166,26 @@ loom_cc_test(
         "//loom/src/loom/ops/func",
         "//loom/src/loom/ops/scalar",
         "//loom/src/loom/target:low_descriptor_registry",
+        "//loom/src/loom/target/test:low_registry",
+        "//loom/src/loom/target/test:lower",
+        "//loom/src/loom/target/test:target_record_table",
+        "//loom/src/loom/util:fact_table",
+        "//runtime/src/iree/testing:gtest",
+        "//runtime/src/iree/testing:gtest_main",
+    ],
+)
+
+loom_cc_test(
+    name = "function_boundary_test",
+    srcs = ["function_boundary_test.cc"],
+    deps = [
+        ":lower",
+        "//loom/src/loom/codegen/low/testing:source_workload",
+        "//loom/src/loom/ir",
+        "//loom/src/loom/ops/func",
+        "//loom/src/loom/ops/low",
+        "//loom/src/loom/target:registers",
+        "//loom/src/loom/target/test:descriptors",
         "//loom/src/loom/target/test:low_registry",
         "//loom/src/loom/target/test:lower",
         "//loom/src/loom/target/test:target_record_table",
